requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
albumentations==1.3.0
apparmor==3.1.1
appdirs==1.4.4
application-utility==1.3.2
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autocommand==2.2.1
base58==2.1.1
beautifulsoup4==4.11.1
bitarray==2.6.0
btrfsutil==5.19.1
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
coloredlogs==15.0.1
commonmark==0.9.1
contourpy==1.0.6
cryptography==38.0.1
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.32
cytoolz==0.12.0
distlib==0.3.6
distro==1.7.0
docopt==0.6.2
docutils==0.19
easydict==1.10
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
exceptiongroup==1.0.4
facenet-pytorch==2.5.2
filelock==3.8.0
fire==0.4.0
Flask==2.2.2
flatbuffers==22.10.26
fonttools==4.38.0
frozenlist==1.3.1
future==0.18.2
gast==0.4.0
gdown==4.5.3
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h11==0.14.0
h5py==3.7.0
hexbytes==0.3.0
humanfriendly==10.0
idna==3.4
imageio==2.22.2
inflect==6.0.0
ipfshttpclient==0.8.0a2
itsdangerous==2.1.2
jaraco.context==4.1.2
jaraco.functools==3.5.2
jaraco.text==3.9.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
keras==2.10.0
Keras-Preprocessing==1.1.2
keyutils==0.6
kiwisolver==1.4.4
LibAppArmor==3.1.1
libclang==14.0.6
lit==14.0.6.dev0
lru-dict==1.1.8
manjaro-sdk==0.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
more-itertools==8.13.0
mpmath==1.2.1
msgpack==1.0.4
mtcnn==0.1.1
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
netsnmp-python==1.0a1
networkx==2.8.7
npyscreen==4.10.5
nspektr==0.4.0
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
onnx==1.12.0
onnxruntime==1.13.1
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
ordered-set==4.1.0
outcome==1.2.0
packaging==21.3
pacman-mirrors==4.23.2
pandas==1.5.1
parsimonious==0.8.1
pep517==0.13.0
Pillow==9.3.0
platformdirs==2.5.2
ply==3.11
prettytable==3.5.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.21.0
pycparser==2.21
pycryptodome==3.15.0
pycups==2.0.1
pycurl==7.45.1
pydantic==1.10.2
pygame==2.1.2
Pygments==2.13.0
PyGObject==3.42.2
pyOpenSSL==22.1.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
pyrsistent==0.18.1
PySide6==6.3.2
pysmbc==1.0.23
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-ffmpeg-video-streaming==0.1.15
pytz==2022.5
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
reportlab==3.6.11
requests==2.28.1
requests-oauthlib==1.3.1
resolvelib==0.8.1
retina-face==0.0.12
rich @ file:///build/python-rich/src/rich-12.6.0/dist/rich-12.6.0-py3-none-any.whl
rlp==2.0.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
selenium==4.6.0
shiboken6==6.3.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sympy==1.11.1
systemd-python==235
TBB==0.2
team==1.0
tenacity==8.1.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
threadpoolctl==3.1.0
tifffile==2022.10.10
tomli==2.0.1
toolz==0.12.0
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
trove-classifiers==2022.9.26
typing_extensions==4.4.0
udiskie==2.4.2
urllib3==1.26.12
validate-pyproject==0.10.1
varint==1.0.2
virtualenv==20.16.0
wcwidth==0.2.5
web3==5.31.0
webencodings==0.5.1
websockets==9.1
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
yarl==1.8.1
youtube-dl==2021.12.17
